refactor: 프로젝트 구조 변경
Application이 지저분해 보였기 때문에 Game -> RacingGame으로 클래스 이름 변경
기존 Application에 있던 코드를 새로운 Game 클래스로 옮기는 식으로 구조 변경
GameValidations -> GameValidator로 이름 변경
